pyproject.toml
[project]
name = "warnet"
version = "0.10.0"
description = "Monitor and analyze the emergent behaviours of bitcoin networks"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["bitcoin", "warnet"]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
]

dependencies = [
    "click==8.1.7",
    "docker==7.1.0",
    "flask==3.0.3",
    "inquirer==3.4.0",
    "kubernetes==30.1.0",
    "rich==13.7.1",
    "tabulate==0.9.0",
    "PyYAML==6.0.2",
]

[project.scripts]
warnet = "warnet.main:cli"

[project.urls]
Homepage = "https://warnet.dev"
GitHub = "https://github.com/bitcoindevproject/warnet"
Pypi = "https://pypi.org/project/warnet/"

[project.optional-dependencies]
build = [
  "twine",
  "build",
]

[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src", "."]
include = ["warnet*", "test_framework*", "resources*"]

[tool.setuptools.package-data]
"resources" = ["**/*"]
